net: Hack to enable IPv6 ECN support
ECN is not currently supported on inet6 sockets.

This is a hack to copy over the ECN bits when ECN-capable transport
has been enabled on the inet_sk socket.

This allows to support DCCPv6 ECN support, but needs to be revised:
 * the inet_sock already has an u8 `tos' field for the DSCP/ECN bits;
 * the same field is present as u8 `tclass' field in ipv6_pinfo;
 * the DS field (RFC 2474) and the ECN bits (RFC 3168) are AF-independent,
   so TOS/Traffic Class could be made AF-independent;
 * sockets that include ipv6_pinfo also include struct inet_sock.

Ideally there should be one field, e.g. `inet->dsfield', but doing
this requires some more careful analysis, since inet_sk(sk)->tos and
inet6_sk(sk)->tclass are referenced all over the source code.

Leaving this for the moment, to be completed later.

Signed-off-by: Gerrit Renker <gerrit@erg.abdn.ac.uk>